berhasil menampilkan absensi

+++ app/src/main/java/com/example/androidkpnew/CallDataAdapterGaji.java
@@ -1,5 +1,5 @@
 package com.example.androidkpnew;
 
 public interface CallDataAdapterGaji {
-    void addToList(String namaPegawai, String rolePegawai, int norekening, float jumlahAbsen);
+    void addToList(String namaPegawai, String rolePegawai, int norekening, Double jumlahAbsen);
 }
